requirements.txt
groq==0.9.0
streamlit==1.31.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.13
faiss-cpu==1.7.4
sentence-transformers==2.3.1
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.2.0
opencv-python==4.9.0.80
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
pandas==2.2.0
numpy==1.26.3
pydantic==2.5.3
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
chromadb==0.4.22
